refactor: 💡 fix typo file name
