+++ src/features/scoreManagementV2/const/ruleOptions.ts
